requirements.txt
# Core (Python 3.9+)
streamlit>=1.30.0
pymupdf>=1.23.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
# spaCy models — install after pip install -r requirements.txt:
#   python -m spacy download en_core_web_lg   (default NER model, English)
#   python -m spacy download en_core_web_sm   (tokeniser for HuggingFace NER models)
#   python -m spacy download de_core_news_lg  (German NER — recommended for Swiss documents)
#   python -m spacy download xx_ent_wiki_sm   (multilingual DE/FR/IT/EN NER)
#
# OCR language packs (required for German/French/Italian scanned documents):
#   brew install tesseract tesseract-lang
cryptography>=42.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# OCR engines
# Surya (recommended — transformer-based, language-agnostic, pure pip)
"surya-ocr<0.5"           # 0.5+ requires Python 3.10+
# Tesseract fallback (requires system binary):
#   brew install tesseract tesseract-lang   (macOS)
#   sudo apt-get install tesseract-ocr tesseract-ocr-deu tesseract-ocr-fra tesseract-ocr-ita

# Optional NER backends (install the ones you need)
flair>=0.13.0
stanza>=1.8.0
transformers>=4.36.0
torch>=2.1.0
# Required by Presidio's HuggingFace/transformers NLP engine
spacy-huggingface-pipelines>=0.0.1

# Optional: local LLM verification (HuggingFace models — Mistral-7B, Phi-3.5-mini, Qwen2.5)
huggingface_hub
accelerate
sentencepiece
protobuf

# Optional: Ollama backend (fastest — requires `brew install ollama` + `ollama serve`)
ollama>=0.2.0